Give the community stat band the same collision floor (#5175)
`.ud-community-band` — the leaderboard's community totals and the
dashboard's cross-city copy of it — lays five tiles out on
`repeat(auto-fit, minmax(150px, 1fr))`, and a grid track has no notion of
what it has to hold. Nothing there stops a value outgrowing its share;
the file's own comment already conceded how tight the budget was
("enough for '11,635 km' on one line at this size, not more").

Unlike #5151 this is not yet visible: the tiles are centred and the
gutter is `--space-lg`, so a value that overhangs by a few px eats into
the gap rather than the next tile, and I could not produce a collision
from any realistic figure — the specs below pass against the old grid
too. It is the guarantee that is missing, not the pixels. What it rests
on instead is a formatting helper in another file: `bigCountMarkup`
abbreviates at a million, and the dashboard's copy doesn't abbreviate at
all (`CrossCityStats.#num` is a bare `toLocaleString`), while
`fmtBigDist` has no abbreviation step at all — so the community distance
grows in its tile without bound.

Switch to flex with a 150px basis. That is the same preferred width the
track floor already named, and the same wrap points, but a flex item's
automatic minimum size floors each tile at its own content, which a track
cannot do — so a value too wide for its share widens its own tile and the
others give up the room.

Equal sizing is kept here deliberately, unlike /about. These tiles are
centred, carry icons, and hold values of similar width, so even columns
read as intended; /about's are left-aligned and range from two glyphs to
nine, where equal columns are exactly what made a full-width value look
jammed against a two-digit neighbour.

Generalize the /about spec to both bands and rename it accordingly. The
community cases pin the contract rather than reproduce a failure, and say
so — they are what would catch a raised font cap, a sixth tile, or a
total that outgrows its abbreviation.
